refactor(caps): require use of Cap in capset!()
BREAKING CHANGE

Previously, using integers in capset!(), such as `capset!(0, 1, 2)`,
would work. It shouldn't; this exposes the library's internals.

Now, the macro internally casts to Cap before casting to u8. Casting
from an integer to an enum is not allowed, so only things such as
capset!(Cap::CHOWN) will work.

In addition, there is a new test to make sure that this works when
passing a capability represented by a variable.