refactor(caps): only store capability *names* in static array
Now, the implementation of FromStr for Cap relies on the fact that the
names in CAP_NAMES are in the same order as the capabilities, instead of
redundantly storing the capability alongside each name.

FromStr is still tested against the derived Debug implementation for
consistency, so this shouldn't introduce any errors.